+++ ecosystem.config.cjs
@@ -1,15 +1,15 @@
 module.exports = {
   apps: [
     {
-      name: 'shiro',
-      script: 'server.js',
+      name: 'Shiro',
+      script: 'npx next start -p 2323',
+      instances: 1,
       autorestart: true,
       watch: false,
-      max_memory_restart: '180M',
+      max_memory_restart: '500M',
       env: {
-        PORT: 2323,
         NODE_ENV: 'production',
-        ...require('dotenv').config().parsed,
+        NEXT_SHARP_PATH: process.env.NEXT_SHARP_PATH,
       },
       log_date_format: 'YYYY-MM-DD HH:mm:ss',
     },