+++ ecosystem.standalone.config.cjs
@@ -1,14 +1,15 @@
 module.exports = {
   apps: [
     {
-      name: 'Shiro',
-      script: 'npx next start -p 2323',
-      instances: 1,
+      name: 'shiro',
+      script: 'server.js',
       autorestart: true,
       watch: false,
-      max_memory_restart: '180M',
+      max_memory_restart: '500M',
       env: {
+        PORT: 2323,
         NODE_ENV: 'production',
+        NEXT_SHARP_PATH: process.env.NEXT_SHARP_PATH,
       },
       log_date_format: 'YYYY-MM-DD HH:mm:ss',
     },