+++ src/app/(app)/notes/[id]/api.tsx
@@ -1,23 +1,33 @@
 import { cache } from 'react'
-import { headers } from 'next/dist/client/components/headers'
 
-import { REQUEST_QUERY } from '~/constants/system'
-import { attachUAAndRealIp } from '~/lib/attach-ua'
+import { attachServerFetch, getAuthFromCookie } from '~/lib/attach-fetch'
 import { getQueryClient } from '~/lib/query-client.server'
 import { requestErrorHandler } from '~/lib/request.server'
 import { queries } from '~/queries/definition'
 
-export const getData = cache(async (params: { id: string }) => {
-  attachUAAndRealIp()
-  const header = headers()
-  const searchParams = new URLSearchParams(header.get(REQUEST_QUERY) || '')
-  const id = params.id
-  const query = queries.note.byNid(
-    id,
-    searchParams.get('password') || undefined,
-  )
-  const data = await getQueryClient()
-    .fetchQuery(query)
-    .catch(requestErrorHandler)
-  return data
-})
+export const getData = cache(
+  async (params: {
+    id: string
+
+    token?: string
+    password?: string
+  }) => {
+    attachServerFetch()
+
+    const { id, password, token } = params
+
+    const query = queries.note.byNid(
+      id,
+      password,
+      token ? `${token}` : undefined,
+    )
+
+    const data = await getQueryClient()
+      .fetchQuery({
+        ...query,
+        staleTime: getAuthFromCookie() ? 0 : undefined,
+      })
+      .catch(requestErrorHandler)
+    return data
+  },
+)