+++ src/components/modules/shared/ToastCard.tsx
@@ -2,8 +2,8 @@
 
 import clsx from 'clsx'
 import { m } from 'framer-motion'
-import type { FC } from 'react'
-import type { ToastProps, TypeOptions } from 'react-toastify/dist/types'
+import type { FC, JSX } from 'react'
+import type { TypeOptions } from 'react-toastify'
 
 import { MotionButtonBase } from '../../ui/button'
 
@@ -19,7 +19,9 @@ const typeMap: Record<TypeOptions, JSX.Element> = {
 
 export const ToastCard: FC<{
   message: string
-  toastProps?: ToastProps
+  toastProps?: {
+    type: TypeOptions
+  }
   iconElement?: JSX.Element
   closeToast?: () => void
   onClick?: () => void
@@ -33,7 +35,7 @@ export const ToastCard: FC<{
       layout="position"
       className={clsx(
         'relative w-full overflow-hidden rounded-xl card-shadow',
-        'my-4 mr-4 px-4 py-5 pr-7',
+        'my-4 mr-4 px-4 py-5 pr-8',
         'bg-zinc-50/90 backdrop-blur-sm dark:bg-neutral-900/90',
         'border border-slate-100/80 dark:border-neutral-900/80',
         'space-x-4',