+++ src/socket/handler.ts
@@ -2,12 +2,14 @@ import { queryClient } from '~/providers/root/react-query-provider'
 import React from 'react'
 import { produce } from 'immer'
 import type {
+  CommentModel,
   NoteModel,
   PaginateResult,
   PostModel,
   RecentlyModel,
   SayModel,
 } from '@mx-space/api-client'
+import type { BusinessEvents } from '@mx-space/webhook'
 import type { InfiniteData } from '@tanstack/react-query'
 import type { ActivityPresence } from '~/models/activity'
 import type { AppRouterInstance } from 'next/dist/shared/lib/app-router-context.shared-runtime'
@@ -43,14 +45,17 @@ import {
   setGlobalCurrentPostData,
 } from '~/providers/post/CurrentPostDataProvider'
 import { queries } from '~/queries/definition'
+import { buildCommentsQueryKey } from '~/queries/keys'
 import { EventTypes } from '~/types/events'
 
-const trackerRealtimeEvent = () => {
+import { WsEvent } from './util'
+
+const trackerRealtimeEvent = (label = 'Socket Realtime Event') => {
   document.dispatchEvent(
     new CustomEvent('impression', {
       detail: {
         action: TrackerAction.Impression,
-        label: 'Socket Realtime Event',
+        label,
       },
     }),
   )
@@ -168,7 +173,7 @@ export const eventHandler = (
     case EventTypes.NOTE_CREATE: {
       const { title, nid } = data as NoteModel
 
-      toast.success('有新的内容发布了：' + `「${title}」`, {
+      toast.success(`有新的内容发布了：「${title}」`, {
         onClick: () => {
           window.peek(`/notes/${nid}`)
         },
@@ -182,7 +187,7 @@ export const eventHandler = (
 
     case EventTypes.POST_CREATE: {
       const { title, category, slug } = data as PostModel
-      toast.success('有新的内容发布了：' + `「${title}」`, {
+      toast.success(`有新的内容发布了：「${title}」`, {
         onClick: () => {
           window.peek(`/posts/${category.slug}/${slug}`)
         },
@@ -250,6 +255,32 @@ export const eventHandler = (
       break
     }
 
+    case EventTypes.COMMENT_CREATE: {
+      const payload = data as {
+        ref: string
+        id: string
+      }
+
+      setTimeout(() => {
+        const queryData = queryClient.getQueryData<
+          InfiniteData<PaginateResult<CommentModel>>
+        >(buildCommentsQueryKey(payload.ref))
+
+        if (!queryData) return
+        for (const page of queryData.pages) {
+          if (page.data.some((comment) => comment.id === payload.id)) {
+            return
+          }
+        }
+
+        queryClient.invalidateQueries({
+          queryKey: buildCommentsQueryKey(payload.ref),
+        })
+      }, 1000)
+
+      break
+    }
+
     case EventTypes.ACTIVITY_LEAVE_PRESENCE: {
       const payload = data as {
         identity: string
@@ -303,23 +334,25 @@ export const eventHandler = (
       break
     }
 
-    case 'shiro#update': {
-      toast.info('站点版本已更新，请刷新页面', {
+    case 'fn#shiro#update': {
+      toast.info('网站已更新，请刷新页面', {
         onClick: () => {
           location.reload()
         },
+        autoClose: false,
       })
+      trackerRealtimeEvent('Shiro Update')
       break
     }
 
     default: {
-      window.dispatchEvent(new CustomEvent(`event:${type}`, { detail: data }))
       if (isDev) {
         // eslint-disable-next-line no-console
-        console.log(type, data)
+        console.info(type, data)
       }
     }
   }
+  WsEvent.emit(type as BusinessEvents, data)
 }
 
 interface ProcessInfo {